Remove unused pde functions
