+++ blockchain/beaconpdeprocess.go
@@ -1,12 +1,8 @@
 package blockchain
 
 import (
-	"encoding/base64"
 	"encoding/json"
-	"math/big"
-	"sort"
 	"strconv"
-	"strings"
 
 	"github.com/incognitochain/incognito-chain/database"
 	"github.com/incognitochain/incognito-chain/database/lvdb"
@@ -51,51 +47,6 @@ func (blockchain *BlockChain) processPDEInstructions(block *BeaconBlock, bd *[]d
 	return nil
 }
 
-func addShareAmountUp(
-	beaconHeight uint64,
-	token1IDStr string,
-	token2IDStr string,
-	contributedTokenIDStr string,
-	contributorAddrStr string,
-	amt uint64,
-	currentPDEState *CurrentPDEState,
-) {
-	pdeShareOnTokenPrefix := string(lvdb.BuildPDESharesKey(beaconHeight, token1IDStr, token2IDStr, contributedTokenIDStr, ""))
-	totalSharesOnToken := uint64(0)
-	for key, value := range currentPDEState.PDEShares {
-		if strings.Contains(key, pdeShareOnTokenPrefix) {
-			totalSharesOnToken += value
-		}
-	}
-	pdeShareKey := string(lvdb.BuildPDESharesKey(beaconHeight, token1IDStr, token2IDStr, contributedTokenIDStr, contributorAddrStr))
-	if totalSharesOnToken == 0 {
-		currentPDEState.PDEShares[pdeShareKey] = amt
-		return
-	}
-	poolPairKey := string(lvdb.BuildPDEPoolForPairKey(beaconHeight, token1IDStr, token2IDStr))
-	poolPair, found := currentPDEState.PDEPoolPairs[poolPairKey]
-	if !found || poolPair == nil {
-		currentPDEState.PDEShares[pdeShareKey] = amt
-		return
-	}
-	poolValue := poolPair.Token1PoolValue
-	if poolPair.Token2IDStr == contributedTokenIDStr {
-		poolValue = poolPair.Token2PoolValue
-	}
-	if poolValue == 0 {
-		currentPDEState.PDEShares[pdeShareKey] = amt
-	}
-	increasingAmt := big.NewInt(0)
-	increasingAmt.Mul(big.NewInt(int64(totalSharesOnToken)), big.NewInt(int64(amt)))
-	increasingAmt.Div(increasingAmt, big.NewInt(int64(poolValue)))
-	currentShare, found := currentPDEState.PDEShares[pdeShareKey]
-	addedUpAmt := increasingAmt.Uint64()
-	if found {
-		addedUpAmt += currentShare
-	}
-	currentPDEState.PDEShares[pdeShareKey] = addedUpAmt
-}
-
 func storePDEPoolForPair(
 	pdePoolForPairKey string,
 	token1IDStr string,
@@ -111,98 +62,6 @@ func storePDEPoolForPair(
 		Token2PoolValue: token2PoolValue,
 	}
 	currentPDEState.PDEPoolPairs[pdePoolForPairKey] = pdePoolForPair
-}
-
-func updateWaitingContributionPairToPool(
-	beaconHeight uint64,
-	waitingContribution1 *lvdb.PDEContribution,
-	waitingContribution2 *lvdb.PDEContribution,
-	currentPDEState *CurrentPDEState,
-) {
-	addShareAmountUp(
-		beaconHeight,
-		waitingContribution1.TokenIDStr,
-		waitingContribution2.TokenIDStr,
-		waitingContribution1.TokenIDStr,
-		waitingContribution1.ContributorAddressStr,
-		waitingContribution1.Amount,
-		currentPDEState,
-	)
-	addShareAmountUp(
-		beaconHeight,
-		waitingContribution1.TokenIDStr,
-		waitingContribution2.TokenIDStr,
-		waitingContribution2.TokenIDStr,
-		waitingContribution2.ContributorAddressStr,
-		waitingContribution2.Amount,
-		currentPDEState,
-	)
-	waitingContributions := []*lvdb.PDEContribution{waitingContribution1, waitingContribution2}
-	sort.Slice(waitingContributions, func(i, j int) bool {
-		return waitingContributions[i].TokenIDStr < waitingContributions[j].TokenIDStr
-	})
-	pdePoolForPairKey := string(lvdb.BuildPDEPoolForPairKey(beaconHeight, waitingContributions[0].TokenIDStr, waitingContributions[1].TokenIDStr))
-	pdePoolForPair, found := currentPDEState.PDEPoolPairs[pdePoolForPairKey]
-	if !found || pdePoolForPair == nil {
-		storePDEPoolForPair(
-			pdePoolForPairKey,
-			waitingContributions[0].TokenIDStr,
-			waitingContributions[0].Amount,
-			waitingContributions[1].TokenIDStr,
-			waitingContributions[1].Amount,
-			currentPDEState,
-		)
-		return
-	}
-	storePDEPoolForPair(
-		pdePoolForPairKey,
-		waitingContributions[0].TokenIDStr,
-		pdePoolForPair.Token1PoolValue+waitingContributions[0].Amount,
-		waitingContributions[1].TokenIDStr,
-		pdePoolForPair.Token2PoolValue+waitingContributions[1].Amount,
-		currentPDEState,
-	)
-}
-
-func contributeToPDE(
-	beaconHeight uint64,
-	pairID string,
-	contributorAddressStr string,
-	tokenIDStr string,
-	contributedAmount uint64,
-	currentPDEState *CurrentPDEState,
-) {
-	waitingContribPairKey := string(lvdb.BuildWaitingPDEContributionKey(beaconHeight, pairID))
-	waitingContribution, found := currentPDEState.WaitingPDEContributions[waitingContribPairKey]
-	if !found || waitingContribution == nil {
-		currentPDEState.WaitingPDEContributions[waitingContribPairKey] = &lvdb.PDEContribution{
-			ContributorAddressStr: contributorAddressStr,
-			TokenIDStr:            tokenIDStr,
-			Amount:                contributedAmount,
-		}
-		return
-	}
-	// there was a waiting pde contribution with the same pairID
-	if tokenIDStr == waitingContribution.TokenIDStr {
-		currentPDEState.WaitingPDEContributions[waitingContribPairKey] = &lvdb.PDEContribution{
-			ContributorAddressStr: contributorAddressStr,
-			TokenIDStr:            tokenIDStr,
-			Amount:                contributedAmount + waitingContribution.Amount,
-		}
-		return
-	}
-	// contributing on the remaining token type of existing pair -> move that pair to pde pool for trading
-	delete(currentPDEState.WaitingPDEContributions, waitingContribPairKey)
-	updateWaitingContributionPairToPool(
-		beaconHeight,
-		&lvdb.PDEContribution{
-			ContributorAddressStr: contributorAddressStr,
-			TokenIDStr:            tokenIDStr,
-			Amount:                contributedAmount,
-		},
-		waitingContribution,
-		currentPDEState,
-	)
 }
 
 func (blockchain *BlockChain) processPDEContributionV2(
@@ -271,42 +130,6 @@ func (blockchain *BlockChain) processPDEContributionV2(
 		)
 		delete(currentPDEState.WaitingPDEContributions, waitingContribPairKey)
 	}
-	return nil
-}
-
-func (blockchain *BlockChain) processPDEContribution(
-	beaconHeight uint64,
-	instruction []string,
-	currentPDEState *CurrentPDEState,
-) error {
-	if currentPDEState == nil {
-		Logger.log.Warn("WARN - [processPDEContribution]: Current PDE state is null.")
-		return nil
-	}
-	if len(instruction) != 4 {
-		return nil // skip the instruction
-	}
-	contentBytes, err := base64.StdEncoding.DecodeString(instruction[3])
-	if err != nil {
-		Logger.log.Errorf("WARNING: an error occured while decoding content string of pde contribution instruction: %+v", err)
-		return nil
-	}
-	var pdeContributionInst metadata.PDEContributionAction
-	err = json.Unmarshal(contentBytes, &pdeContributionInst)
-	if err != nil {
-		Logger.log.Errorf("WARNING: an error occured while unmarshaling pde contribution instruction: %+v", err)
-		return nil
-	}
-	meta := pdeContributionInst.Meta
-	pairID := meta.PDEContributionPairID
-	contributeToPDE(
-		beaconHeight,
-		pairID,
-		meta.ContributorAddressStr,
-		meta.TokenIDStr,
-		meta.ContributedAmount,
-		currentPDEState,
-	)
 	return nil
 }
