+++ blockchain/pdeutils.go
@@ -4,6 +4,7 @@ import (
 	"encoding/binary"
 	"encoding/json"
 	"fmt"
+	"sort"
 	"strings"
 
 	"github.com/incognitochain/incognito-chain/database"
@@ -194,4 +195,47 @@ func storePDEStateToDB(
 		return err
 	}
 	return nil
+}
+
+func updateWaitingContributionPairToPoolV2(
+	beaconHeight uint64,
+	waitingContribution1 *lvdb.PDEContribution,
+	waitingContribution2 *lvdb.PDEContribution,
+	currentPDEState *CurrentPDEState,
+) {
+	addShareAmountUpV2(
+		beaconHeight,
+		waitingContribution1.TokenIDStr,
+		waitingContribution2.TokenIDStr,
+		waitingContribution1.TokenIDStr,
+		waitingContribution1.ContributorAddressStr,
+		waitingContribution1.Amount,
+		currentPDEState,
+	)
+
+	waitingContributions := []*lvdb.PDEContribution{waitingContribution1, waitingContribution2}
+	sort.Slice(waitingContributions, func(i, j int) bool {
+		return waitingContributions[i].TokenIDStr < waitingContributions[j].TokenIDStr
+	})
+	pdePoolForPairKey := string(lvdb.BuildPDEPoolForPairKey(beaconHeight, waitingContributions[0].TokenIDStr, waitingContributions[1].TokenIDStr))
+	pdePoolForPair, found := currentPDEState.PDEPoolPairs[pdePoolForPairKey]
+	if !found || pdePoolForPair == nil {
+		storePDEPoolForPair(
+			pdePoolForPairKey,
+			waitingContributions[0].TokenIDStr,
+			waitingContributions[0].Amount,
+			waitingContributions[1].TokenIDStr,
+			waitingContributions[1].Amount,
+			currentPDEState,
+		)
+		return
+	}
+	storePDEPoolForPair(
+		pdePoolForPairKey,
+		waitingContributions[0].TokenIDStr,
+		pdePoolForPair.Token1PoolValue+waitingContributions[0].Amount,
+		waitingContributions[1].TokenIDStr,
+		pdePoolForPair.Token2PoolValue+waitingContributions[1].Amount,
+		currentPDEState,
+	)
 }
